+++ src/main/java/jcrapi2/api/intern/cards/Card.java
@@ -34,4 +34,7 @@ public class Card {
 
     @SerializedName("iconUrls")
     private IconUrls iconUrls;
+
+    @SerializedName("maxEvolutionLevel")
+    private Integer maxEvolutionLevel;
 }