+++ src/main/java/jcrapi2/api/intern/players/battlelog/Card.java
@@ -40,4 +40,10 @@ public class Card {
 
     @SerializedName("iconUrls")
     private IconUrls iconUrls;
+
+    @SerializedName("maxEvolutionLevel")
+    private Integer maxEvolutionLevel;
+
+    @SerializedName("evolutionLevel")
+    private Integer evolutionLevel;
 }