+++ src/main/java/jcrapi2/api/intern/players/battlelog/Team.java
@@ -57,4 +57,7 @@ public class Team {
 
     @SerializedName("rounds")
     private List<Round> rounds;
+
+    @SerializedName("globalRank")
+    private Integer globalRank;
 }