+++ src/main/java/jcrapi2/api/intern/players/info/CurrentDeck.java
@@ -43,4 +43,7 @@ public class CurrentDeck {
 
     @SerializedName("iconUrls")
     private IconUrls iconUrls;
+
+    @SerializedName("maxEvolutionLevel")
+    private Integer maxEvolutionLevel;
 }